+++ src/main/java/roomescape/reservation/service/CreateReservationService.java
@@ -1,0 +1,84 @@
+package roomescape.reservation.service;
+
+import org.springframework.stereotype.Service;
+import roomescape.auth.dto.LoginMember;
+import roomescape.common.exception.AlreadyInUseException;
+import roomescape.common.exception.EntityNotFoundException;
+import roomescape.member.domain.Member;
+import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.ReservationTime;
+import roomescape.reservation.dto.request.ReservationCreateRequest;
+import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.repository.ReservationRepository;
+import roomescape.reservation.repository.ReservationTimeRepository;
+import roomescape.theme.domain.Theme;
+import roomescape.theme.repository.ThemeRepository;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+
+@Service
+public class CreateReservationService {
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
+    private final ThemeRepository themeRepository;
+    private final MemberRepository memberRepository;
+
+    public CreateReservationService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository, MemberRepository memberRepository) {
+        this.reservationRepository = reservationRepository;
+        this.reservationTimeRepository = reservationTimeRepository;
+        this.themeRepository = themeRepository;
+        this.memberRepository = memberRepository;
+    }
+
+    public ReservationResponse create(final ReservationCreateRequest request) {
+        if (isAlreadyBooked(request)) {
+            throw new AlreadyInUseException("reservation is already in use");
+        }
+
+        Reservation reservation = createReservation(request);
+        LocalDateTime now = LocalDateTime.now();
+        validateDateTime(now, reservation.getDate(), reservation.getTime().getStartAt());
+
+        Reservation savedReservation = reservationRepository.save(reservation);
+
+        return ReservationResponse.from(savedReservation);
+    }
+
+    private boolean isAlreadyBooked(final ReservationCreateRequest request) {
+        return reservationRepository.existsByDateAndTimeIdAndThemeId(
+                request.date(), request.timeId(), request.themeId()
+        );
+    }
+
+    private Reservation createReservation(final ReservationCreateRequest request) {
+        ReservationTime reservationTime = getReservationTime(request);
+        Theme theme = getTheme(request);
+        LoginMember loginMember = request.loginMember();
+        Member member = memberRepository.findById(loginMember.id())
+                .orElseThrow(() -> new EntityNotFoundException("등록되지 않은 회원입니다."));
+        return new Reservation(member, request.date(), reservationTime, theme);
+    }
+
+    private Theme getTheme(final ReservationCreateRequest request) {
+        Long themeId = request.themeId();
+        return themeRepository.findById(themeId)
+                .orElseThrow(() -> new EntityNotFoundException("theme not found id =" + themeId));
+    }
+
+    private ReservationTime getReservationTime(final ReservationCreateRequest request) {
+        Long timeId = request.timeId();
+        return reservationTimeRepository.findById(timeId)
+                .orElseThrow(() -> new EntityNotFoundException("reservationsTime not found id =" + timeId));
+    }
+
+    private void validateDateTime(final LocalDateTime now, final LocalDate date, final LocalTime time) {
+        LocalDateTime dateTime = LocalDateTime.of(date, time);
+
+        if (now.isAfter(dateTime)) {
+            throw new IllegalArgumentException("이미 지난 예약 시간입니다.");
+        }
+    }
+}
