+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -40,8 +40,6 @@ import roomescape.theme.repository.ThemeRepository;
 @Import(ReservationService.class)
 class ReservationServiceTest {
 
-    private final LocalDateTime now = LocalDateTime.now();
-
     @Autowired
     private ReservationRepository reservationRepository;
     @Autowired
@@ -80,122 +78,6 @@ class ReservationServiceTest {
         List<ReservationResponse> result = reservationService.getAll();
 
         assertThat(result).isEmpty();
-    }
-
-    @DisplayName("예약을 추가한다.")
-    @Test
-    void test3() {
-        // given
-        Theme savedTheme = themeRepository.save(new Theme("포스티", "공포", "wwww.um.com"));
-        Long themeId = savedTheme.getId();
-
-        LocalTime time = LocalTime.of(8, 0);
-        ReservationTime savedTime = reservationTimeRepository.save(new ReservationTime(time));
-        Long timeId = savedTime.getId();
-        Member savedMember = memberRepository.save(new Member("포스티", "test@test.com", "12341234", Role.MEMBER));
-
-        LocalDate date = nextDay();
-
-        ReservationCreateRequest requestDto =
-                new ReservationCreateRequest(date, timeId, themeId, LoginMember.of(savedMember));
-
-        // when
-        ReservationResponse result = reservationService.create(requestDto);
-
-        // then
-        SoftAssertions softAssertions = new SoftAssertions();
-
-        softAssertions.assertThat(result.member().name()).isEqualTo("포스티");
-        softAssertions.assertThat(result.date()).isEqualTo(date);
-        softAssertions.assertThat(result.time()).isEqualTo(new ReservationTimeResponse(timeId, time));
-        softAssertions.assertThat(result.theme())
-                .isEqualTo(new ThemeResponse(themeId, savedTheme.getName(), savedTheme.getDescription(),
-                        savedTheme.getThumbnail()));
-
-        softAssertions.assertAll();
-    }
-
-    @DisplayName("이미 존재하는 예약과 동일하면 예외가 발생한다.")
-    @Test
-    void test4() {
-        // given
-        Theme savedTheme = themeRepository.save(new Theme("포스티", "공포", "wwww.um.com"));
-        Long themeId = savedTheme.getId();
-
-        LocalTime time = LocalTime.of(8, 0);
-        ReservationTime savedTime = reservationTimeRepository.save(new ReservationTime(time));
-        Long timeId = savedTime.getId();
-        Member member = new Member("포스티", "test@test.com", "12341234", Role.MEMBER);
-        Member savedMember = memberRepository.save(member);
-        LocalDate date = nextDay();
-
-        ReservationCreateRequest requestDto =
-                new ReservationCreateRequest(date, timeId, themeId, LoginMember.of(savedMember));
-        reservationService.create(requestDto);
-
-        // when & then
-        assertThatThrownBy(() -> reservationService.create(requestDto))
-                .isInstanceOf(AlreadyInUseException.class);
-    }
-
-    @DisplayName("과거 날짜에 예약을 추가하면 예외가 발생한다.")
-    @Test
-    void test5() {
-        // given
-        Theme savedTheme = themeRepository.save(new Theme("포스티", "공포", "wwww.um.com"));
-        Long themeId = savedTheme.getId();
-        Member member = new Member("포스티", "test@test.com", "12341234", Role.MEMBER);
-        Member savedMember = memberRepository.save(member);
-
-        LocalDate date = now.toLocalDate();
-        LocalTime pastTime = now.toLocalTime().minusMinutes(1);
-
-        ReservationTime savedTime = reservationTimeRepository.save(new ReservationTime(pastTime));
-        Long timeId = savedTime.getId();
-
-        ReservationCreateRequest requestDto =
-                new ReservationCreateRequest(date, timeId, themeId, LoginMember.of(savedMember));
-
-        // when & then
-        assertThatThrownBy(() -> reservationService.create(requestDto))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("존재하지 않는 예약 시간 ID로 저장하면 예외를 반환한다.")
-    @Test
-    void test6() {
-        Theme savedTheme = themeRepository.save(new Theme("포스티", "공포", "wwww.um.com"));
-        Long themeId = savedTheme.getId();
-        Member member = new Member("포스티", "test@test.com", "12341234", Role.MEMBER);
-        Member savedMember = memberRepository.save(member);
-
-        LocalDate date = nextDay();
-
-        Long notExistId = 1000L;
-        ReservationCreateRequest requestDto =
-                new ReservationCreateRequest(date, notExistId, themeId, LoginMember.of(savedMember));
-
-        assertThatThrownBy(() -> reservationService.create(requestDto))
-                .isInstanceOf(EntityNotFoundException.class);
-    }
-
-    @DisplayName("존재하지 않는 테마 ID로 저장하면 예외를 반환한다.")
-    @Test
-    void notExistThemeId() {
-        LocalDate date = now.toLocalDate().plusDays(1);
-
-        LocalTime time = LocalTime.of(8, 0);
-        ReservationTime savedTime = reservationTimeRepository.save(new ReservationTime(time));
-        Long timeId = savedTime.getId();
-        Member member = new Member("포스티", "test@test.com", "12341234", Role.MEMBER);
-        Member savedMember = memberRepository.save(member);
-
-        Long notExistId = 1000L;
-        ReservationCreateRequest requestDto =
-                new ReservationCreateRequest(date, timeId, notExistId, LoginMember.of(savedMember));
-
-        assertThatThrownBy(() -> reservationService.create(requestDto))
-                .isInstanceOf(EntityNotFoundException.class);
     }
 
     @DisplayName("예약을 삭제한다")
